C-Language Basics

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	long long int mylong_int;
 	float myfloat;
 
-	printf("The size of a char is: %lu byte(s)\n", sizeof(mychar));
-	printf("The size of an int is: %lu byte(s)\n", sizeof(myInt));
-	printf("The size of a long int is: %lu byte(s)\n", sizeof(mylongint));
-	printf("The size of a long long int is: %lu byte(s)\n", sizeof(mylong_int));
-	printf("The size of a float is: %lu byte(s)\n", sizeof(myfloat));
+	printf("The size of a char is: %lu byte(s)\n", (unsigned long)sizeof(mychar));
+	printf("The size of an int is: %lu byte(s)\n", (unsigned long)sizeof(myInt));
+	printf("The size of a long int is: %lu byte(s)\n", (unsigned long)sizeof(mylongint));
+	printf("The size of a long long int is: %lu byte(s)\n", (unsigned long)sizeof(mylong_int));
+	printf("The size of a float is: %lu byte(s)\n", (unsigned long)sizeof(myfloat));
 	return (0);
 }
